+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Web: load profile picture [#5053](https://github.com/rokwire/illinois-app/issues/5053).
 
 ## [7.0.17] - 2025-04-24
 ### Added